👾  Fix thread execution for setting wallpaper
We were getting ANR (app not responding) error reports that straced to
`wallpaperManager.setBitmap`.

The Glide stuff needed to be run on the main thread in order to process
and animate the image (I think?), so the solution to this seemed to be
that the `setBitmap` call should be executed in a background thread.
This is possible by using `AsyncTask`.

This seems to work fine in the simulator so 🤞